+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@
   * Ticket https://github.com/PandaHugMonster/py-simputils-config/issues/38
   * Structure of tests is adjusted to replicate the issue with tests
   * This is a quick fix through a small hack of pre-importing `ConfigHub` for the project.
-    It should be resolved eventually during refactorings.
+    The hack should be refactored eventually.
 * Added `none_considered_empty` (default `False`) argument for `ConfigStore`, that considers any `None` value
   for parsing as "undefined", which prevents overriding previous aggregated config value with `None`.
   Heavily useful in combination with `argparser`, because `None` values from `argparser` might
@@ -16,6 +16,10 @@
   * [example-config-from-env-vars.py](../examples/example-config-from-env-vars.py)
 * Fixed a bug with applied confs history, when argparser's `Namespace` is displayed as `ConfigStore` instead of `args`
   when `ConfigHub` is used.
+* Added alias-property `history` for `applied_confs` of `ConfigStore`
+* Fixed a small bug with incorrect `ConfigStoreType` used when `ConfigStore` merged, causing some
+  merged history items to be mis-represented as `dict` instead of correct `ConfigStore` type.
+* 
 
 ## 1.1.0